octeontx2-pf: Vary completion queue event size
Completion Queue Entry(CQE) is a descriptor written
by hardware to notify software about the send and
receive completion status. The CQE can be of size
128 or 512 bytes. A 512 bytes CQE can hold more receive
fragments pointers compared to 128 bytes CQE. This
patch enables to modify CQE size using:
<ethtool -G cqe-size N>.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/71680
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>

+++ drivers/net/ethernet/marvell/octeontx2/nic/otx2_common.h
@@ -180,8 +180,7 @@ struct otx2_hw {
 
 #define OTX2_DEFAULT_RBUF_LEN	2048
 	u16			rbuf_len;
-
-	u16			xqe_size;
+	u32			xqe_size;
 
 	/* NPA */
 	u32			stack_pg_ptrs;  /* No of ptrs per stack page */
